Home page constriction
I have started writing the main page of the website. It is not yet, complete, for I still have to write the pages to the listed categories.
I'm still not building css, because I still need to decide how I plan to visually organise pages. This will come later, once the skeleton is built